+++ actix-http/CHANGES.md
@@ -1,6 +1,9 @@
 # Changes
 
 ## Unreleased - 2022-xx-xx
+### Changed
+- Minimum supported Rust version (MSRV) is now 1.59 due to transitive `time` dependency.
+
 ### Fixed
 - Avoid possibility of dispatcher getting stuck while back-pressuring I/O. [#2369]